TechNest/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.5
amqp==5.3.1
anyio==4.10.0
asgiref==3.9.0
attrs==25.3.0
banal==1.0.6
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dataset==1.6.2
datasets==4.0.0
dill==0.3.8
Django==5.0.6
django-cors-headers==4.7.0
django-oauth-toolkit==3.0.1
django-unfold==0.55.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
drf-yasg==1.21.10
filelock==3.19.1
firebase_admin==7.1.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
inflection==0.5.1
Jinja2==3.1.6
joblib==1.5.2
jwcrypto==1.5.6
kiwisolver==1.4.9
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.6
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mysqlclient==2.2.7
Naked==0.1.32
networkx==3.5
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
PyJWT==2.10.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==6.4.0
regex==2025.9.1
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentencepiece==0.2.1
setuptools==80.9.0
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlparse==0.5.3
stripe==12.5.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0+cu128
torchaudio==2.8.0+cu128
torchvision==0.23.0+cu128
tqdm==4.67.1
transformers==4.56.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
vncorenlp==1.0.3
wcwidth==0.2.13
wheel==0.45.1
xxhash==3.5.0
yarl==1.20.1
